+++ management/src/commands/index.ts
@@ -7,3 +7,4 @@ export * from "./remove-identity-from-user";
 export * from "./send-identity-validation-code";
 export * from "./update-connect-application";
 export * from "./update-identity-from-user";
+export * from "./send-2fa-verification-code";