+++ management/tests/commands/index.test.ts
@@ -8,6 +8,7 @@ import {
   sendIdentityValidationCode,
   updateProviderApplication,
   updateIdentityFromUser,
+  send2FaVerificationCode,
 } from "../../index";
 
 describe("Commands", () => {
@@ -21,5 +22,6 @@ describe("Commands", () => {
     expect(sendIdentityValidationCode).toBeInstanceOf(Function);
     expect(updateProviderApplication).toBeInstanceOf(Function);
     expect(updateIdentityFromUser).toBeInstanceOf(Function);
+    expect(send2FaVerificationCode).toBeInstanceOf(Function);
   });
 });